+++ tox.ini
@@ -46,6 +46,7 @@ deps =
     mypy
     pep8-naming
     pydocstyle>=2.10
+    PyJWT
     pylint
     pyproject-flake8<6.0.0
     pytest
@@ -102,6 +103,7 @@ commands =
 description = Run integration tests
 deps =
     juju
+    PyJWT
     pytest
     pytest-asyncio
     pytest-operator
